extract session UI
