+++ examples/todomvc/web/src/lib.rs
@@ -1,4 +1,3 @@
-use anyhow::anyhow;
 use graphql_client::{GraphQLQuery, Response};
 use indexmap::IndexMap;
 use seed::{prelude::*, *};
@@ -8,13 +7,13 @@ use uuid::Uuid;
 use web_sys::HtmlInputElement;
 
 mod auth;
+mod session;
 
 const ENTER_KEY: u32 = 13;
 const ESC_KEY: u32 = 27;
 const API_URL: &str = "http://localhost:3030/graphql";
 const ACTIVE: &str = "active";
 const COMPLETED: &str = "completed";
-const STORAGE_KEY: &str = "todomvc-session";
 
 type TodoId = Uuid;
 
@@ -49,7 +48,7 @@ where
         .await
 }
 
-struct Model {
+pub struct Model {
     base_url: Url,
     data: Data,
     refs: Refs,
@@ -117,7 +116,7 @@ impl TodoFilter {
     }
 }
 
-enum Msg {
+pub enum Msg {
     TodosFetched(fetch::Result<Response<get_todos::ResponseData>>),
     UrlChanged(subs::UrlChanged),
 
@@ -141,9 +140,7 @@ enum Msg {
     EditingTodoSaved(fetch::Result<Response<update_todo::ResponseData>>),
     CancelTodoEdit,
 
-    Login,
-    LoggedIn(Option<auth::Claims>),
-    Logout,
+    Session(session::Msg),
 }
 
 fn update(msg: Msg, model: &mut Model, orders: &mut impl Orders<Msg>) {
@@ -326,30 +323,7 @@ fn update(msg: Msg, model: &mut Model, orders: &mut impl Orders<Msg>) {
             data.editing_todo = None;
         }
 
-        Login => {
-            if let Ok(auth_url) = auth::get_login_url() {
-                window()
-                    .open_with_url_and_target(auth_url.to_string().as_str(), "_self")
-                    .expect("couldn't open window");
-            }
-            orders.skip();
-        }
-        LoggedIn(Some(claims)) => {
-            data.user = Some(claims.name);
-            Url::new()
-                .set_path(model.base_url.hash_path())
-                .go_and_replace();
-        }
-        LoggedIn(None) => {
-            data.user = None;
-        }
-
-        Logout => {
-            if let Ok(()) = auth::logout() {
-                LocalStorage::remove(STORAGE_KEY).expect("problem removing key from local storage");
-                data.user = None;
-            }
-        }
+        Session(msg) => session::update(msg, model, orders),
     }
 }
 
@@ -376,7 +350,7 @@ fn view(model: &Model) -> impl IntoNodes<Msg> {
 fn view_header(new_todo_title: &str, user: &Option<String>) -> Node<Msg> {
     header![
         C!["header"],
-        view_nav(user),
+        session::view_nav(user),
         h1!["todos"],
         input![
             C!["new-todo"],
@@ -390,32 +364,6 @@ fn view_header(new_todo_title: &str, user: &Option<String>) -> Node<Msg> {
             }),
             input_ev(Ev::Input, Msg::NewTodoTitleChanged),
         ]
-    ]
-}
-
-fn view_nav(user: &Option<String>) -> Node<Msg> {
-    nav![
-        C!["auth-text"],
-        if let Some(user) = user {
-            nodes![
-                span![format!("{} ", user)],
-                a![
-                    C!["auth-link"],
-                    mouse_ev(Ev::Click, |_| Msg::Logout),
-                    "logout"
-                ]
-            ]
-        } else {
-            nodes![
-                span!["Please "],
-                a![
-                    C!["auth-link"],
-                    mouse_ev(Ev::Click, |_| Msg::Login),
-                    "login"
-                ],
-                span![" to modify todos"]
-            ]
-        },
     ]
 }
 
@@ -574,32 +522,8 @@ fn view_clear_completed(todos: &IndexMap<TodoId, Todo>) -> Option<Node<Msg>> {
     })
 }
 
-fn get_claims(url: &Url) -> anyhow::Result<Option<auth::Claims>> {
-    if let Ok(auth_response) = LocalStorage::get(STORAGE_KEY) {
-        let claims = auth::get_claims(&auth_response)?;
-        return Ok(Some(claims));
-    } else {
-        let url = url.to_string();
-        let url = url
-            .strip_prefix("/#")
-            .ok_or_else(|| anyhow!("url doesn't start with \"/#\""))?;
-        if let Ok(auth_response) = serde_urlencoded::from_str::<auth::AuthResponse>(url) {
-            let claims = auth::get_claims(&auth_response)?;
-            LocalStorage::insert(STORAGE_KEY, &auth_response)
-                .map_err(|e| anyhow!("cannot insert into localstorage {:?}", e))?;
-            return Ok(Some(claims));
-        }
-    }
-    Ok(None)
-}
-
 fn after_mount(url: Url, orders: &mut impl Orders<Msg>) -> AfterMount<Model> {
-    match get_claims(&url) {
-        Ok(claims) => {
-            orders.perform_cmd(async move { Msg::LoggedIn(claims) });
-        }
-        Err(e) => error!(e),
-    };
+    session::after_mount(&url, orders);
     orders.perform_cmd(async {
         let request = GetTodos::build_query(get_todos::Variables);
         let response = send_graphql_request(&request).await;
